move Options package into pkg directory so it can be used by other go modules. Currently kusk (previously kgw cli) needs access to the options (#309)
