+++ pkg/spec/extension_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/getkin/kin-openapi/openapi3"
 	"github.com/stretchr/testify/require"
 
-	"github.com/kubeshop/kusk-gateway/internal/options"
+	"github.com/kubeshop/kusk-gateway/pkg/options"
 )
 
 func TestGetOptions(t *testing.T) {
